requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
beautifulsoup4==4.6.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
click-completion==0.4.1
colorama==0.3.9
crayons==0.1.2
cryptography==2.3.1
decorator==4.3.0
Flask==0.11.1
Flask-Cache==0.13.1
Flask-Cors==3.0.6
flask-mongoengine==0.8.2
Flask-RESTful==0.3.5
Flask-WTF==0.13.1
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
line-profiler==2.1.2
Markdown==2.6.11
MarkupSafe==1.0
marshmallow==2.15.4
mongoengine==0.15.0
more-itertools==4.3.0
parso==0.3.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.4
Pycco==0.6.0
pycparser==2.18
pycycle==0.0.8
pyDes==2.0.1
Pygments==2.2.0
pymongo==3.7.0
PyMySQL==0.9.2
pyprof2calltree==1.4.3
pystache==0.5.4
pytest==3.7.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
robotframework==3.0.4
scandir==1.9.0
shellingham==1.2.4
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
smartypants==2.0.1
SQLAlchemy==1.2.11
traitlets==4.3.2
typing==3.6.4
urllib3==1.23
wcwidth==0.1.7
webargs==4.0.0
Werkzeug==0.14.1
WTForms==2.2.1
yappi==0.98
